requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Web UI
flask==3.0.0
flask-cors==4.0.0

# Kubernetes & Container Management
kubernetes==29.0.0
docker==7.0.0
grpcio==1.60.0
grpcio-tools==1.60.0

# Database & Caching
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1
asyncpg==0.29.0

# Machine Learning & Prediction
torch==2.1.2
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
statsmodels==0.14.1
xgboost==2.0.3
lightgbm==4.2.0

# Deep Learning
transformers==4.36.2
pytorch-lightning==2.1.3
tensorboard==2.15.1

# Time Series
prophet==1.1.5
pmdarima==2.0.4
tslearn==0.6.3

# Optimization
scipy==1.11.4
cvxpy==1.4.2
pulp==2.8.0
networkx==3.2.1

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
py-ebpf==0.1.0
psutil==5.9.7

# API & Serialization
httpx==0.26.0
aiohttp==3.9.1
msgpack==1.0.7
protobuf==4.25.2

# Task Queue & Async
celery==5.3.4
celery[redis]==5.3.4
aiokafka==0.10.0
aioredis==2.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
typer==0.9.0
rich==13.7.0
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
locust==2.20.0
faker==22.0.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pylint==3.0.3
bandit==1.7.6

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.4
mkdocstrings[python]==0.24.0

# Distributed Computing
ray==2.9.1
dask[complete]==2024.1.0

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Deployment
gunicorn==21.2.0
supervisor==4.2.5

# Data Processing
pyarrow==14.0.2
h5py==3.10.0
tables==3.9.2

# Graph & Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Scientific Computing
joblib==1.3.2
numba==0.59.0
cython==3.0.8
